fix : episode put operation : when setting duration over short's max value
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@4271 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/web/api/ApiContent.java
@@ -1322,9 +1322,9 @@ public class ApiContent extends ApiGeneric {
         // duration
         String durationStr = req.getParameter("duration");
         if (durationStr != null) {
-            Short duration = null;
+            Integer duration = null;
             try {
-                duration = Short.valueOf(durationStr);
+                duration = Integer.valueOf(durationStr);
             } catch (NumberFormatException e) {
             }
             if ((duration == null) || (duration <= 0)) {
